requirements.txt
google-adk
litellm
openai
groq
azure-search-documents
azure-identity
azure-storage-blob
python-dotenv
requests
duckduckgo-search
langchain
langchain-community
mcp
fastapi
uvicorn[standard]
pydantic
# Speech-to-text
SpeechRecognition
pydub
